refactor(fun): separate entertainment feature services
